requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai 
python-dotenv
scholarly
scrape-google-scholar-py